+++ src/constants/recruitData.tsx
@@ -1,0 +1,176 @@
+import Link from "next/link";
+import type { ReactNode } from "react";
+
+export const categories = ["전체", "선발", "지원", "활동"];
+
+export type QnAItem = {
+  field: string;
+  question: string;
+  answer: ReactNode[];
+};
+
+export const qnaList: QnAItem[] = [
+  {
+    field: "선발",
+    question: "모집시기가 언제인가요?",
+    answer: [
+      "큐시즘은 상반기, 하반기 총 2번 신규 운영진 및 학회원을 모집해요.",
+      "운영진 리쿠르팅 이후 학회원 리쿠르팅이 시작되니",
+      "모집시기를 놓치고 싶지 않다면 홈페이지 상단에 있는 배너를 클릭하여 미리 알림을 받아보세요.",
+    ],
+  },
+  {
+    field: "선발",
+    question: "경쟁률이 어떻게 되나요?",
+    answer: ["매 기수 모집마다 경쟁률이 상이하게 나타나요. 따라서 경쟁률은 말씀드리기 어려운 점 양해부탁드려요."],
+  },
+  {
+    field: "활동",
+    question: "토요일 활동 시간대가 어떻게 되나요?",
+    answer: ["정규 세션은 보통 12~6시 사이에 진행돼요. 따라서 토요일 오후 시간에는 큐시즘 활동이 가능해야 해요!"],
+  },
+  {
+    field: "지원",
+    question: "졸업예정 학부생도 지원이 가능한가요?",
+    answer: ["큐시즘은 한 기수동안 재학생, 휴학생, 졸업 유예자처럼 대학생 신분을 유지하는 사람들만 지원이 가능해요."],
+  },
+  {
+    field: "활동",
+    question: "학회 수료시, 수료증을 받을 수 있나요?",
+    answer: ["네, 학회의 마지막 커리큘럼에서 서류 기준을 충족한 학회원에게 수료증을 제공하고 있어요."],
+  },
+  {
+    field: "활동",
+    question: "정기 커리큘럼에 불참하게 되면 어떻게 되나요?",
+    answer: [
+      "정기 커리큘럼에 불참하면 정관에 따라 벌점이 부과돼요.",
+      "최대 벌점을 초과하면 수료를 못하게 될 수 있어요. 정기 커리큘럼에는 최대한 참여해야 수료에 지장이 없어요.",
+    ],
+  },
+  {
+    field: "활동",
+    question: "프로젝트는 어떻게 진행되나요?",
+    answer: [
+      "매 기수마다 차이가 있지만, 밋업데이 이후 꾸준히 프로젝트를 이어나가는 팀이 존재해요.",
+      <>
+        자세한 이야기는{" "}
+        <Link href="/projects/meetup" className="text-[#0055FF]">
+          프로젝트 페이지
+        </Link>
+        에서 확인해보세요.
+      </>,
+    ],
+  },
+  {
+    field: "지원",
+    question: "운영진과 학회원의 차이가 무엇인가요?",
+    answer: [
+      "운영진은 학회원의 일원으로써 일반 학회원과 동일하게 모든 행사에 참여해요.",
+      "다만, 운영진은 큐시즘 행사 전반에 대해 관리하고 운영하는 역할을 하며 더욱 큐시즘을 알차게 보낼 수 있게 해줘요!",
+    ],
+  },
+  {
+    field: "지원",
+    question: "큐시즘이 처음인데 운영진부터 지원해도 되나요?",
+    answer: [
+      "네 물론이에요 :)",
+      "큐시즘 활동을 하면서 더 많은 경험을 쌓고 알차게 보내고 싶은 분이라면 언제나 환영이에요.",
+      "매 기수 신규 운영진으로 들어와 함께 하는 분들이 많기 때문에",
+      "큐시즘에 대한 열정만 있으시다면 전혀 부담 가지실 필요 없어요!",
+    ],
+  },
+  {
+    field: "선발",
+    question: "운영진 지원 후 불합격이면, 일반 학회원으로 재신청해도 불이익은 없나요?",
+    answer: [
+      "네, 전혀 불이익 없어요!",
+      "모든 지원 서류는 모집이 끝난 후 즉시 폐기하므로 학회원으로 재신청하셔도 전혀 문제되지 않아요.",
+    ],
+  },
+  {
+    field: "선발",
+    question: "서류/면접 전형은 어떻게 진행되나요?",
+    answer: ["서류는 위더스 웹을 통해 받고있으며 면접은 오프라인 대면 면접으로 진행하고 있어요."],
+  },
+  {
+    field: "선발",
+    question: "한 기수에 몇 명 정도 뽑나요?",
+    answer: ["한 기수에 64명이 활동하며, 활동연장한 OB를 제외한 인원을 뽑아 기수별로 상이해요."],
+  },
+  {
+    field: "지원",
+    question: "전공자가 아니어도 지원 가능한가요?",
+    answer: ["네, 전공이 아니더라도 큐시즘에 대한 열정만 있다면 참여할 수 있어요!"],
+  },
+  {
+    field: "지원",
+    question: "이전 기수에서 떨어졌는데 다시 지원해도 되나요?",
+    answer: [
+      "네, 재지원 가능해요.",
+      "모든 지원 서류는 모집이 끝난 후 즉시 폐기하므로 학회원으로 재신청하셔도 전혀 문제되지 않아요.",
+    ],
+  },
+  {
+    field: "활동",
+    question: "정규 세션 말고도 추가 스터디가 있나요?",
+    answer: [
+      "네, 정규 세션 외에 자율적으로 스터디를 열어 공부할 수 있어요.",
+      "성장하고 싶은 주제라면 모두 가능해요!",
+    ],
+  },
+];
+
+export const recruitInfoItems = [
+  {
+    role: "기획",
+    activity: [
+      "문제를 빠르게 정의하고, 사용자 관점에서 실현 가능한 기능으로 정리해요.",
+      "각 파트가 이해할 수 있도록 화면 흐름, 예외 케이스, 우선순위를 문서로 남겨요.",
+      "진행 중에 바뀌는 요구를 파악해서 디자인·개발과 조율하며 팀 방향을 맞춰요.",
+    ],
+    people: [
+      "데이터나 리서치를 근거로 가설을 세우고 논리적으로 설명할 수 있는 분",
+      "디자이너·개발자와의 협업 속에서 자연스럽게 우선순위를 판단할 수 있는 분",
+      "변화하는 상황 속에서도 책임감 있게 추진하면서 함께 일하는 과정을 즐길 수 있는 분",
+    ],
+  },
+  {
+    role: "디자인",
+    activity: [
+      "기획 의도를 이해해 UX 플로우를 만들고, 실제 화면(UI)로 구체화해요.",
+      "개발이 구현하기 쉬운 형태로 컴포넌트와 가이드를 정리해요.",
+      "피드백을 받아 사용성을 높이고 완성도를 올리는 수정을 진행해요.",
+    ],
+    people: [
+      "자신의 디자인을 논리적으로 설명하고, 실제 사용자 경험에 적용해본 분",
+      "포트폴리오 안에서 본인 기여도를 명확하게 정리할 수 있는 분",
+      "협업 과정에서 긍정적으로 소통하고, 스스로를 객관적으로 보며 성장하려는 태도를 가진 분",
+    ],
+  },
+  {
+    role: "프론트",
+    activity: [
+      "웹의 동작 원리를 바탕으로 디자인 시안을 실제 화면으로 구현해요.",
+      "API 명세서를 보고 백엔드와 기능을 조율하거나 오류 원인을 함께 찾아내요.",
+      "디자인 시스템이나 공통 UI를 프론트 단에서 재사용 가능하게 구조화해요.",
+    ],
+    people: [
+      "핵심 웹 기술에 대한 기본기를 갖추고, 다른 파트와 소통하며 기능을 맞춰본 분",
+      "PR과 코드리뷰로 타인의 코드를 읽고 피드백한 경험이 있는 분",
+      "프로젝트 규모에 맞춰 기술 스택과 아키텍처를 직접 고민해보고 싶은 분",
+    ],
+  },
+  {
+    role: "백엔드",
+    activity: [
+      "웹 서비스 구조와 서버 동작을 이해하고, 요구사항에 맞는 API와 데이터 구조를 설계해요.",
+      "프론트엔드와 API 명세를 기준으로 소통하며 오류 원인을 함께 분석해요.",
+      "서비스의 비즈니스 로직과 유지보수성을 고려해 안정적인 시스템을 구현해요.",
+    ],
+    people: [
+      "주요 백엔드 기술의 기본기를 갖추고 있고, 구조를 이해하려는 습관이 있는 분",
+      "API 문서를 보고 프론트엔드와 요구를 조율하거나 문제를 풀어본 경험이 있는 분",
+      "팀 상황에 따라 기술 스택을 선택하고 더 나은 대안을 고민할 수 있는 분",
+    ],
+  },
+];
